--- Editor/EngageAnimationEvents/LocalShootPosition.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 08d4b3214979493283f983e3d01cb90c
-timeCreated: 1739081720
