+++ Editor/EngageAnimationEvents/Vec3Types/LocalShootPosition.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3fa1a405bac0f4ca5ab03fa5fe462d61
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
